+++ src/main/modules/i18n/locales/cy.json
@@ -3307,7 +3307,7 @@
       "WHAT_SUPPORT": "What support do you, experts or witnesses need?",
       "FOR_EXAMPLE": "For example, a witness requires a courtroom with step-free access."
     },
-    "NOTICES_AND_ORDERS": {
+    "ORDERS_AND_NOTICES": {
       "CCJ_REQUEST": "County Court Judgment request",
       "CLAIMANT_DQ": "Claimant Directions Questionnaire",
       "CLAIMANT_RESPONSE_RECEIPT": "Claimant response receipt",
