+++ src/main/views/features/dashboard/orders-and-notices.njk
@@ -50,7 +50,7 @@
 {% if courtDocuments.documents.length > 0 %}
     {% set courtDocRows = (courtDocRows.push([
         {
-            text: t('PAGES.NOTICES_AND_ORDERS.COURT_DOCUMENTS'),
+            text: t('PAGES.ORDERS_AND_NOTICES.COURT_DOCUMENTS'),
             classes: "govuk-body govuk-!-font-weight-bold"
         },
         {
